Entirely new Generator based API.
This change eliminates Inputs from Config (seperation
of concerns), replaces the Translate method with a
Generator class on which FindFiles can be called for
each input directory to be included and the code
generated by invoking WriteTo.

It should now be possible to do incremental builds,
which might be of use to someone.