Files should not implement sort.Interface.
filepath.Walk already walks the directory in deterministically.

The only real use of sort.Sort(Files(...)) was when joining
multiple FindFiles outputs together. Files can still be sorted
with sort.Slice if needed. Otherwise just ensure that each Files
is joined in the same order.

This changes the order of go-bindata files. The order is now
determined by the order of input file paths.